+++ packages/ui/src/Components/Input/Select.tsx
@@ -14,7 +14,7 @@ import { getTextValue, TextValue } from '../../textParser'
 import useHover from '../../hooks/events/useHover'
 import { Text } from '../Text'
 import useDropdown, { OnSelect } from '../../hooks/overlay/useDropdown'
-import useScopedState from '../../hooks/useInputValue'
+import useScopedState from '../../hooks/useScopedState'
 import renderChildren from '../../util/renderChildren'
 
 type SelectInputProps = {
@@ -65,7 +65,7 @@ export const Select: FunctionComponent<SelectInputProps> = ({
 
   const [stateValue, setValue] = useScopedState<
     DropdownOption | DropdownOption[]
-  >(value, identifier, isFocus)
+  >(value, identifier)
 
   /**
    * Force state-update when value is updated externally.
